Use regular expressions to generate the prompt.
The name of the ad group can sometimes contain a lot more info than is needed, and sometimes we want to extract parts of it, e.g. a location.

This change will allow you to use named capturing groups in regular expressions, and combine those in the prompt, allowing for more complex prompt generation.

For example if an adgroup is named `EN_UK_London_generic` you could use `^(?<language>[a-zA-Z]{2})_(?<country>[a-zA-Z]+)_(?<city>[a-zA-Z]+)` to capture the various fields, and create a prompt by `A photo of a landmark in ${city}, ${country}.